Update to 2.0.0

+++ neonbot/classes/embed.py
@@ -6,6 +6,7 @@ from typing import Any, Optional
 import discord
 from discord.ext import commands
 
+from .view import Button, View
 from ..helpers.constants import CHOICES_EMOJI, PAGINATION_EMOJI
 
 
@@ -17,7 +18,7 @@ class Embed(discord.Embed):
             super().__init__(**kwargs)
         self.color = 0x59ABE3
 
-    def add_field(self, name: Any, value: Any, *, inline: bool = True) -> None:
+    def add_field(self, name: Any, value: Any, *, inline: bool = True) -> Embed:
         super().add_field(name=name, value=value, inline=inline)
         return self
 
@@ -27,22 +28,22 @@ class Embed(discord.Embed):
         url: str = discord.Embed.Empty,
         *,
         icon_url: str = discord.Embed.Empty,
-    ) -> None:
+    ) -> Embed:
         super().set_author(name=name, url=url, icon_url=icon_url)
         return self
 
     def set_footer(
         self, text: str = discord.Embed.Empty, *, icon_url: str = discord.Embed.Empty
-    ) -> None:
+    ) -> Embed:
         super().set_footer(text=text, icon_url=icon_url)
         return self
 
-    def set_image(self, url: str) -> None:
+    def set_image(self, url: str) -> Embed:
         if url:
             super().set_image(url=url)
         return self
 
-    def set_thumbnail(self, url: Optional[str]) -> None:
+    def set_thumbnail(self, url: Optional[str]) -> Embed:
         if url:
             super().set_thumbnail(url=url)
         return self
@@ -72,7 +73,7 @@ class PaginationEmbed:
         self.index = 0
         self.embed = Embed()
 
-        self.msg: discord.Message = None
+        self.msg: Optional[discord.Message] = None
 
     async def build(self) -> None:
         self.authorized_users.append(self.ctx.author.id)
@@ -80,17 +81,18 @@ class PaginationEmbed:
         await self._send()
 
         if len(self.embeds) > 1:
-            asyncio.gather(self._add_reactions(), self._listen())
+            await asyncio.gather(self._add_reactions(), self._listen())
 
     async def _send(self) -> None:
         embed = self.embed.copy()
         embed.description = self.embeds[self.index].description
 
         if len(self.embeds) > 1:
-            embed.description += f"\n\n**Page {self.index+1}/{len(self.embeds)}**"
+            embed.description += f"\n\n**Page {self.index + 1}/{len(self.embeds)}**"
 
         if self.msg:
-            return await self.msg.edit(embed=embed)
+            await self.msg.edit(embed=embed)
+            return
 
         self.msg = await self.ctx.send(embed=embed)
 
@@ -99,13 +101,13 @@ class PaginationEmbed:
 
         msg = self.msg
 
-        def check(reaction: discord.Reaction, user: discord.User) -> bool:
-            if not user.bot and reaction.emoji != "🗑":
-                asyncio.ensure_future(reaction.remove(user))
+        def check(r: discord.Reaction, user: discord.User) -> bool:
+            if not user.bot and r.emoji != "🗑":
+                asyncio.ensure_future(r.remove(user))
             return (
-                reaction.emoji in PAGINATION_EMOJI
+                r.emoji in PAGINATION_EMOJI
                 and user.id in self.authorized_users
-                and reaction.message.id == msg.id
+                and r.message.id == msg.id
             )
 
         try:
@@ -141,8 +143,7 @@ class PaginationEmbed:
             if m.content.isdigit():
                 self.bot.loop.create_task(self.bot.delete_message(m))
                 if (
-                    int(m.content) >= 0
-                    and int(m.content) <= len(self.embeds)
+                    0 <= int(m.content) <= len(self.embeds)
                     and m.channel.id == self.ctx.channel.id
                 ):
                     return True
@@ -188,41 +189,43 @@ class EmbedChoices:
             await self.ctx.send(embed=Embed("Empty choices."), delete_after=5)
             return self
 
-        await self._send_choices()
-        asyncio.ensure_future(self._react())
-        await self._listen()
+        await self.send_choices()
 
         return self
 
-    async def _send_choices(self) -> None:
+    async def send_choices(self) -> None:
         embed = Embed(title=f"Choose 1-{len(self.entries)} below.")
 
         for index, entry in enumerate(self.entries, start=1):
-            embed.add_field(f"{index}. {entry.title}", entry.url, inline=False)
+            embed.add_field(f"{index}. {entry['title']}", entry['url'], inline=False)
 
-        self.msg = await self.ctx.send(embed=embed)
+        buttons = self.get_buttons()
 
-    async def _listen(self) -> None:
-        try:
-            reaction, _ = await self.ctx.bot.wait_for(
-                "reaction_add",
-                timeout=10,
-                check=lambda reaction, user: reaction.emoji in CHOICES_EMOJI
-                and self.ctx.author == user
-                and reaction.message.id == self.msg.id,
-            )
-            if reaction.emoji == "🗑":
-                raise asyncio.TimeoutError
-        except asyncio.TimeoutError:
-            self.value = -1
-        else:
-            self.value = CHOICES_EMOJI.index(reaction.emoji)
-        finally:
+        self.msg = await self.ctx.send(embed=embed, view=buttons)
+
+        await buttons.wait()
+
+    def get_buttons(self) -> discord.ui.View:
+        async def callback(button: discord.ui.Button, interaction: discord.Interaction):
+            if interaction.user != self.ctx.author:
+                return
+
+            if button.emoji and button.emoji.name == CHOICES_EMOJI[-1]:
+                self.value = -1
+            else:
+                self.value = int(button.label)
+
+            button.view.stop()
             await self.bot.delete_message(self.msg)
 
-    async def _react(self) -> None:
-        for emoji in CHOICES_EMOJI[0 : len(self.entries)] + [CHOICES_EMOJI[-1]]:
-            try:
-                await self.msg.add_reaction(emoji)
-            except discord.NotFound:
-                break
+
+        buttons = [
+            {"label": 1},
+            {"label": 2},
+            {"label": 3},
+            {"label": 4},
+            {"label": 5},
+            {"emoji": CHOICES_EMOJI[-1]},
+        ]
+
+        return View.create_button(buttons, callback)
